+++ addons/l10n_in/__manifest__.py
@@ -37,6 +37,7 @@ Sheet, now only Vertical format has been permitted Which is Supported By Odoo.
         'views/res_config_settings_views.xml',
         'views/product_template_view.xml',
         'views/port_code_views.xml',
+        'views/res_company_views.xml',
         'views/report_invoice.xml',
         'views/res_country_state_view.xml',
         'views/res_partner_views.xml',
