+++ addons/l10n_in/models/company.py
@@ -1,8 +1,9 @@
-from odoo import api, models
-
+from odoo import api, fields, models
 
 class ResCompany(models.Model):
     _inherit = 'res.company'
+
+    l10n_in_upi_id = fields.Char(string="UPI Id")
 
     @api.model_create_multi
     def create(self, vals_list):
