+++ addons/l10n_in/views/res_company_views.xml
