--- addons/l10n_in_upi/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import models
