--- addons/l10n_in_upi/models/res_company.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import fields, models
-
-
-class Company(models.Model):
-    _inherit = 'res.company'
-
-    l10n_in_upi_id = fields.Char(string="UPI Id")
